requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.0.0+cpu
torchvision==0.15.1+cpu
Pillow
matplotlib
scikit-image
opencv-python
